+++ infra/mongo/mongo_handler.go
@@ -1,4 +1,4 @@
-package infra
+package mongo
 
 import (
 	mgo "gopkg.in/mgo.v2"
@@ -6,25 +6,25 @@ import (
 	"github.com/hirokisan/go-sample-clean-architecture/interface/database"
 )
 
-type MongoHandler struct {
+type Handler struct {
 	Session    *mgo.Session
 	Db         string
 	Collection string
 }
 
-func NewMongoHandler(url string, db string, col string) database.MongoHandler {
+func NewHandler(url string, db string, col string) database.MongoHandler {
 	session, err := mgo.Dial(url)
 	if err != nil {
 		panic(err)
 	}
-	mongoHandler := new(MongoHandler)
-	mongoHandler.Session = session
-	mongoHandler.Db = db
-	mongoHandler.Collection = col
-	return mongoHandler
+	Handler := new(Handler)
+	Handler.Session = session
+	Handler.Db = db
+	Handler.Collection = col
+	return Handler
 }
 
-func (m *MongoHandler) Insert(dest interface{}) error {
+func (m *Handler) Insert(dest interface{}) error {
 	err := m.Session.DB(m.Db).C(m.Collection).Insert(dest)
 	if err != nil {
 		return err
@@ -32,7 +32,7 @@ func (m *MongoHandler) Insert(dest interface{}) error {
 	return nil
 }
 
-func (m *MongoHandler) FindAll(res interface{}) error {
+func (m *Handler) FindAll(res interface{}) error {
 	err := m.Session.DB(m.Db).C(m.Collection).Find(nil).All(res)
 	if err != nil {
 		return err
